time_core::convert::sealed

Trait MultipleOf

pub trait MultipleOf<T, Output> {
    const VALUE: Output;
}

A trait for defining the ratio of two units of time.

This trait is used to implement the per method on the various structs.

Required Associated Constants

const VALUE: Output

The number of one unit of time in the other.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.