Trait time_core::convert::sealed::Per

pub trait Per<T> {
    type Output;

    const VALUE: Self::Output;
}

A trait for defining the ratio of two units of time.

This trait is used to implement the per method on the various structs.

Required Associated Types

type Output

The smallest unsigned integer type that can represent VALUE.

Required Associated Constants

const VALUE: Self::Output

The number of one unit of time in the other.