time_core::convert

Struct Second

pub struct Second;

A unit of time representing exactly one second.

Implementations

impl Second

pub const fn per<T>(_larger: T) -> <T as DefaultOutput<Self>>::Output
where T: MultipleOf<Self, T::Output> + DefaultOutput<Self> + Copy,

Obtain the number of times Second can fit into T. If T is smaller than Second, the code will fail to compile. The return type is the smallest unsigned integer type that can represent the value.

Valid calls:

Second::per(Second) (returns u8)

Second::per(Minute) (returns u8)

Second::per(Hour) (returns u16)

Second::per(Day) (returns u32)

Second::per(Week) (returns u32)

pub const fn per_t<Output>( larger: impl MultipleOf<Self, Output> + Copy, ) -> Output

Obtain the number of times Second can fit into T. If T is smaller than Second, the code will fail to compile. The return type is any primitive numeric type that can represent the value.

Valid calls:

Second::per(Second) (returns u8, u16, u32, u64, u128, usize, i8, i16, i32, i64, i128, isize, f32, or f64)

Second::per(Minute) (returns u8, u16, u32, u64, u128, usize, i8, i16, i32, i64, i128, isize, f32, or f64)

Second::per(Hour) (returns u16, u32, u64, u128, usize, i16, i32, i64, i128, isize, f32, or f64)

Second::per(Day) (returns u32, u64, u128, usize, i32, i64, i128, isize, f32, or f64)

Second::per(Week) (returns u32, u64, u128, usize, i32, i64, i128, isize, f32, or f64)