time_core::unit

Struct Day

pub struct Day<const N: u128 = 1>;

A unit of time representing exactly N days.

Implementations

impl Day<1>

pub const fn per<T>(_larger: T) -> <T as DefaultOutput<Self>>::Output
where T: MultipleOf<Self, T::Output> + DefaultOutput<Self> + Copy,

Obtain the number of times Day can fit into T. If T is smaller than Day, the code will fail to compile. The return type is the smallest unsigned integer type that can represent the value.

Valid calls:

Day::per(Day) (returns u8)

Day::per(Week) (returns u8)

pub const fn per_t<Output>( larger: impl MultipleOf<Self, Output> + Copy, ) -> Output

Obtain the number of times Day can fit into T. If T is smaller than Day, the code will fail to compile. The return type is any primitive numeric type that can represent the value.

Valid calls:

Day::per(Day) (returns u8, u16, u32, u64, u128, usize, i8, i16, i32, i64, i128, isize, f32, or f64)

Day::per(Week) (returns u8, u16, u32, u64, u128, usize, i8, i16, i32, i64, i128, isize, f32, or f64)