time_core
Module util

Utility functions.

Functions

days_in_year

Get the number of calendar days in a given year.

is_leap_year

Returns if the provided year is a leap year in the proleptic Gregorian calendar. Uses astronomical year numbering.

weeks_in_year

Get the number of weeks in the ISO year.