time_core::util

Function days_in_month

pub const fn days_in_month(month: u8, year: i32) -> u8

Get the number of days in the month of a given year.

assert_eq!(days_in_month(2, 2020), 29);

Note: This function is not exposed by the time crate. It is an implementation detail.