time_core::util

Module range_validated

Versions of functions that are optimized for when the year has already been range-validated.

The implementations of these functions depend on whether the large-dates feature is enabled.

Note: This module is not exposed by the time crate. It is an implementation detail.

Functions

days_in_month

Get the number of days in the month of a given year, assuming the year has already been range-validated.

days_in_year

Get the number of calendar days in a given year, assuming the year has already been range-validated.

is_leap_year

Returns if the provided year is a leap year in the proleptic Gregorian calendar, assuming the year has already been range-validated.